requirements.txt
asgiref==3.4.1
Django==4.0
django-appconf==1.0.5
django-ckeditor==6.2.0
django-imagekit==4.1.0
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-sass==1.1.0
libsass==0.21.0
pilkit==2.0
Pillow==8.4.0
psycopg2-binary==2.9.2
six==1.16.0
sqlparse==0.4.2
